Refacto prepare_diagnosis
