+++ app/jobs/prepare_solicitation_diagnosis_job.rb
@@ -1,6 +1,6 @@
 class PrepareSolicitationDiagnosisJob < ApplicationJob
   def perform(solicitation_id)
     solicitation = Solicitation.find(solicitation_id)
-    solicitation.prepare_diagnosis(nil)
+    CreateDiagnosis::PrepareDiagnosis.new(solicitation, nil).call
   end
 end
