+++ app/models/concerns/diagnosis_creation.rb
@@ -4,88 +4,6 @@ module DiagnosisCreation
   # The creation params hash for a diagnosis has nested attributes for #facility and #facility#company.
   # These will be used `fields_for` form helpers.
 
-  module SolicitationMethods
-    # Some preconditions can be verified without actually trying to create the Diagnosis
-    def may_prepare_diagnosis?
-      self.preselected_subject.present? &&
-        FormatSiret.siret_is_valid(FormatSiret.clean_siret(self.siret))
-    end
-
-    # Attempt to create a diagnosis up to the last step with the information from the solicitation.
-    # Use the preselected attributes (subject & institution) to create the needs and matches.
-    #
-    # returns nil and sets self.prepare_diagnosis_errors on error.
-    # returns the diagnosis on success
-    def prepare_diagnosis(advisor)
-      return unless may_prepare_diagnosis?
-
-      prepare_diagnosis_errors = nil
-      diagnosis = nil
-      Diagnosis.transaction do
-        # Step 0: create with the facility
-        diagnosis = CreateDiagnosis::CreateOrUpdateDiagnosis.new(
-          {
-            advisor: advisor,
-            solicitation: self,
-            facility_attributes: computed_facility_attributes
-          }, self.diagnosis
-        ).call
-
-        diagnosis.autofill_steps
-
-        # Rollback on error!
-        if diagnosis.errors.present?
-          prepare_diagnosis_errors = diagnosis.errors
-          diagnosis = nil
-          raise ActiveRecord::Rollback
-        end
-      end
-
-      # Save or clear the error
-      self.update(prepare_diagnosis_errors: prepare_diagnosis_errors, diagnosis: diagnosis)
-      diagnosis
-    end
-
-    def computed_facility_attributes
-      if self.with_siret?
-        { siret: FormatSiret.clean_siret(self.siret) }
-      else
-        {
-          insee_code: retrieve_insee_code,
-          company_attributes: { name: self.full_name }
-        }
-      end
-    end
-
-    def retrieve_insee_code
-      # TODO : à revoir quand on aura une meilleure gestion des zones
-      query = location.parameterize
-      Api::Adresse::SearchMunicipality.new(query).call[:insee_code]
-    end
-
-    ## Store ActiveModel::Errors details as json…
-    #
-    def prepare_diagnosis_errors=(diagnosis_errors)
-      self.prepare_diagnosis_errors_details = diagnosis_errors&.details
-    end
-
-    ## … and build a temporary Diagnosis to recreate Errors.
-    # This lets us call use the errors in the UI just like regular ActiveModel errors.
-    def prepare_diagnosis_errors
-      diagnosis_errors = Diagnosis.new.errors
-
-      self.prepare_diagnosis_errors_details&.each do |attr, errors|
-        errors.each { |h| h.each_value { |error| diagnosis_errors.add(attr, error.to_sym) } }
-      end
-
-      diagnosis_errors
-    end
-
-    def with_siret?
-      self.siret.present?
-    end
-  end
-
   module DiagnosisMethods
     def prepare_needs_from_solicitation
       return unless solicitation.present? && needs.blank?
